requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
crispy-bootstrap5==2024.10
cryptography==44.0.2
cssselect2==0.7.0
cycler==0.12.1
Django==5.1.4
django-allauth==65.7.0
django-crispy-forms==2.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.55.3
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sounddevice==0.5.1
sqlparse==0.5.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
tf_keras==2.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
weasyprint==63.1
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zopfli==0.2.3.post1
